+++ starsky/starsky/clientapp/src/contexts/sorter.spec.ts
@@ -46,8 +46,8 @@ describe("sorter", () => {
 
     const resultList = sorter(list, SortType.imageFormat);
 
-    expect(resultList[0].fileName).toBe("b");
-    expect(resultList[1].fileName).toBe("a");
+    expect(resultList[0].fileName).toBe("a");
+    expect(resultList[1].fileName).toBe("b");
   });
 
   it("sort on imageFormat", () => {
@@ -67,6 +67,42 @@ describe("sorter", () => {
 
     expect(resultList[0].fileName).toBe("b");
     expect(resultList[1].fileName).toBe("a");
+  });
+
+  it("sort on imageFormat 2 same type", () => {
+    const list = [
+      {
+        fileName: "a",
+        imageFormat: ImageFormat.bmp
+      } as IFileIndexItem,
+      {
+        fileName: "b",
+        imageFormat: ImageFormat.bmp
+      } as IFileIndexItem
+    ] as IFileIndexItem[];
+
+    const resultList = sorter(list, SortType.imageFormat);
+
+    expect(resultList[0].fileName).toBe("a");
+    expect(resultList[1].fileName).toBe("b");
+  });
+
+  it("sort on imageFormat 2 same type 1", () => {
+    const list = [
+      {
+        fileName: "b",
+        imageFormat: ImageFormat.unknown
+      } as IFileIndexItem,
+      {
+        fileName: "a",
+        imageFormat: ImageFormat.unknown
+      } as IFileIndexItem
+    ] as IFileIndexItem[];
+
+    const resultList = sorter(list, SortType.imageFormat);
+
+    expect(resultList[0].fileName).toBe("a");
+    expect(resultList[1].fileName).toBe("b");
   });
 
   it("sort on imageFormat, example 2", () => {