+++ starsky/starsky/clientapp/src/contexts/sorter.ts
@@ -19,7 +19,8 @@ export function sorter(
 
         const enumOrder = Object.values(ImageFormat);
         return (
-          enumOrder.indexOf(a.imageFormat) - enumOrder.indexOf(b.imageFormat)
+          enumOrder.indexOf(a.imageFormat) - enumOrder.indexOf(b.imageFormat) ||
+          a.fileName.localeCompare(b.fileName, "en", { sensitivity: "base" })
         );
       });
     default: